// admin/src/app/core/api/model/productDTO.ts
/**
 * Univers des Âmes API
 * API REST pour la boutique et les services de l\'Univers des Âmes
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StoneDTO } from './stoneDTO';
import { ProductSizeDTO } from './productSizeDTO';
import { ProductVariantDTO } from './productVariantDTO';


export interface ProductDTO { 
    /**
     * ID unique du produit
     */
    id: number;
    name: string;
    /**
     * Description détaillée en HTML ou texte brut
     */
    description: string;
    price: number;
    /**
     * Liste des pierres associées
     */
    stones?: Array<StoneDTO>;
    imageUrl?: string;
    type: ProductDTO.TypeEnum;
    /**
     * Nombre de séances (pour les soins/coaching)
     */
    sessionCount?: number;
    /**
     * Durée de validité en mois
     */
    durationMonths?: number;
    /**
     * Liste des variantes associées à un service (prix, durée, nb de séances..)
     */
    variants?: Array<ProductVariantDTO>;
    /**
     * Liste des tailles associées et disponibles pour un produit (S, M, L...)
     */
    sizes?: Array<ProductSizeDTO>;
    /**
     * Temps de repos entre deux rendez-vous (minutes)
     */
    bufferTimeMinutes?: number;
    isSubscription: boolean;
    recurringInterval?: ProductDTO.RecurringIntervalEnum;
}
export namespace ProductDTO {
    export type TypeEnum = 'PHYSICAL' | 'ENERGY_CARE' | 'CARD_READING' | 'COACHING';
    export const TypeEnum = {
        Physical: 'PHYSICAL' as TypeEnum,
        EnergyCare: 'ENERGY_CARE' as TypeEnum,
        CardReading: 'CARD_READING' as TypeEnum,
        Coaching: 'COACHING' as TypeEnum
    };
    export type RecurringIntervalEnum = 'month' | 'year' | 'week';
    export const RecurringIntervalEnum = {
        Month: 'month' as RecurringIntervalEnum,
        Year: 'year' as RecurringIntervalEnum,
        Week: 'week' as RecurringIntervalEnum
    };
}